backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-telegram-bot==20.7
openpyxl==3.1.2
aiohttp==3.9.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
